Moved binary threshold and quantize to filters. Renamed quantize normalized to quantize.
