+++ gtf_impl/grading.py
@@ -1,4 +1,1 @@
-import torch
-
-
 # https://filmicworlds.com/blog/minimal-color-grading-tools/
