+++ gtf_nodes/convert.py
@@ -141,46 +141,3 @@ class ChannelMax:
     def f(gtf: torch.Tensor) -> tuple[torch.Tensor]:
         tensor_max = gtf_max(gtf, (0))
         return (tensor_max, )
-
-
-class BinaryThreshold:
-    @staticmethod
-    def INPUT_TYPES():
-        return {
-            "required": {
-                "gtf": ("GTF", ),
-                "threshold": ("FLOAT", {"default": 0.5, "step": 0.001}),
-            }
-        }
-
-    RETURN_TYPES = ("GTF", )
-    RETURN_NAMES = ("gtf", )
-    CATEGORY = "gtf/convert"
-    FUNCTION = "f"
-
-    @staticmethod
-    def f(gtf: torch.Tensor, threshold: float) -> tuple[torch.Tensor]:
-        thresholded = (gtf >= threshold).to(torch.float)
-        return (thresholded, )
-
-
-class QuantizeNormalized:
-    @staticmethod
-    def INPUT_TYPES():
-        return {
-            "required": {
-                "gtf": ("GTF", ),
-                "steps": ("INT", {"default": 256, "min": 2, "max": 1_000_000}),
-                "mode": ([*C.F_MAP.keys()], ),
-            }
-        }
-
-    RETURN_TYPES = ("GTF", )
-    RETURN_NAMES = ("gtf", )
-    CATEGORY = "gtf/convert"
-    FUNCTION = "f"
-
-    @staticmethod
-    def f(gtf: torch.Tensor, steps: int, mode: str) -> tuple[torch.Tensor]:
-        quantized = C.quantize_normalized(gtf, steps, mode)
-        return (quantized, )
